refactor record reader into class
