+++ backend/services.plr-intake/.vscode/settings.json
@@ -15,5 +15,6 @@
     "**/bin": true,
     "**/logs": true,
     "**/obj": true,
+    "**/Data/Migrations": true
   },
 }
